+++ src/main/java/com/marsspiders/ukwa/solr/SolrSearchService.java
@@ -11,28 +11,19 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.net.URI;
-import java.net.URISyntaxException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 import static com.marsspiders.ukwa.controllers.CollectionController.TYPE_COLLECTION;
-import static com.marsspiders.ukwa.controllers.CollectionController.TYPE_TARGET;
 import static com.marsspiders.ukwa.solr.AccessToEnum.VIEWABLE_ANYWHERE;
 import static com.marsspiders.ukwa.util.SolrUtil.escapeQueryChars;
 import static com.marsspiders.ukwa.util.SolrUtil.toEncoded;
 import static com.marsspiders.ukwa.util.SolrUtil.toMultipleConditionsQuery;
+import static com.marsspiders.ukwa.util.SolrUtil.toMultipleConditionsQueryWithPreCondition;
 import static java.lang.String.format;
-import static java.util.stream.Collectors.toList;
 
 @Service
 public class SolrSearchService {
@@ -57,6 +48,7 @@ public class SolrSearchService {
     private static final String FIELD_TITLE = "title";
     private static final String FIELD_URL = "url";
     private static final String FIELD_WAYBACK_DATE = "wayback_date";
+    private static final String FIELD_COLLECTION = "collection";
 
     private static final String EXCLUDE_FACET_TAG_NAME = "filter";
     private static final String EXCLUDE_MARKER_TAG = "{!tag=" + EXCLUDE_FACET_TAG_NAME + "}";
@@ -64,10 +56,6 @@ public class SolrSearchService {
     private static final String DATE_PART_AFTER_YEAR = "-01-01T00:00:00Z";
     private static final int FACET_RANGE_START_YEAR = 1980;
     private static final int FACET_RANGE_END_YEAR = 2040;
-
-    private final Map<String, List<String>> collectionIdsToHosts = new HashMap<>();
-    private final Map<String, String> urlsToCollectionId = new HashMap<>();
-    private final Map<String, String> idsToNames = new HashMap<>();
 
     @Autowired
     SolrCommunicator communicator;
@@ -119,81 +107,6 @@ public class SolrSearchService {
         return sendRequest(queryString, CollectionInfo.class, rowsLimit, startFrom);
     }
 
-    public Map<String, String> getParentCollections(List<String> hostsPairs) {
-        Map<String, String> collectionIdsToName = new HashMap<>();
-
-        if (hostsPairs == null || hostsPairs.size() == 0) {
-            return collectionIdsToName;
-        }
-
-        List<String> actualHosts = toActualHosts(hostsPairs);
-        determineUrlsByCollections();
-
-        log.info("Getting parent collections by hosts: " + actualHosts);
-
-        for (String url : urlsToCollectionId.keySet()) {
-            if(url == null){
-               continue;
-            }
-
-            String domainName = toDomainName(url);
-
-            String collectionId = urlsToCollectionId.get(url);
-            if (actualHosts.contains(domainName) && collectionIdsToName.get(collectionId) == null) {
-                collectionIdsToName.put(collectionId, idsToNames.get(collectionId));
-            }
-        }
-
-        log.info("Parent collections by hosts are: " + collectionIdsToName);
-
-
-        return collectionIdsToName;
-    }
-
-    private void determineUrlsByCollections() {
-        if (urlsToCollectionId.size() > 0) {
-            return;
-        }
-
-        log.info("Determining urls of all collections");
-
-        SolrSearchResult<CollectionInfo> allCollections = fetchAllCollections();
-        allCollections.getResponseBody().getDocuments()
-                .stream()
-                .forEach(d -> idsToNames.put(d.getId(), d.getName()));
-
-        SolrSearchResult<CollectionInfo> childCollections = fetchChildCollections(idsToNames.keySet(), TYPE_TARGET, 100000, 0);
-        childCollections.getResponseBody().getDocuments()
-                .stream()
-                .forEach(d -> {
-                    String url = d.getUrl();
-                    String collectionId = d.getParentId();
-                    urlsToCollectionId.put(url, collectionId);
-
-                    List<String> hosts = collectionIdsToHosts.get(collectionId);
-                    if (hosts == null) {
-                        hosts = new ArrayList<>();
-                    }
-
-                    if(url != null){
-                        hosts.add(toDomainName(url));
-                    }
-
-                    collectionIdsToHosts.put(collectionId, hosts);
-                });
-    }
-
-    private static String toDomainName(String url) {
-        try {
-            URI uri = new URI(url);
-            String domain = uri.getHost();
-            return domain.startsWith("www.") ? domain.substring(4) : domain;
-        } catch (URISyntaxException e) {
-            log.error(format("Failed to build URI from url '%s'", url), e);
-            return null;
-        }
-    }
-
     public SolrSearchResult<ContentInfo> searchContent(SearchByEnum searchLocation,
                                                        String textToSearch,
                                                        long rowsLimit,
@@ -206,7 +119,7 @@ public class SolrSearchService {
                                                        Date fromDatePicked,
                                                        Date toDatePicked,
                                                        List<String> rangeDates,
-                                                       List<String> collectionIds) {
+                                                       List<String> collections) {
         log.info("Searching content for '" + textToSearch + "' by " + searchLocation);
 
         String accessToValue = accessTo == null
@@ -218,21 +131,22 @@ public class SolrSearchService {
         String dateQuery = generateDateQuery(fromDatePicked, toDatePicked, rangeDates);
 
         String contentTypeNotEmpty = FIELD_CONTENT_TYPE_NORM + ":['' TO *]";
-        String contentTypesConditionQuery = toMultipleConditionsQuery(contentTypes, FIELD_CONTENT_TYPE_NORM);
+        String contentTypesConditionQuery = toMultipleConditionsQueryWithPreCondition(contentTypes, FIELD_CONTENT_TYPE_NORM);
         String contentTypeQuery = EXCLUDE_MARKER_TAG + contentTypeNotEmpty + contentTypesConditionQuery;
 
         String publicSuffixesNotEmpty = FIELD_PUBLIC_SUFFIX + ":['' TO *]";
-        String publicSuffixesConditionQuery = toMultipleConditionsQuery(publicSuffixes, FIELD_PUBLIC_SUFFIX);
+        String publicSuffixesConditionQuery = toMultipleConditionsQueryWithPreCondition(publicSuffixes, FIELD_PUBLIC_SUFFIX);
         String publicSuffixesQuery = EXCLUDE_MARKER_TAG + publicSuffixesNotEmpty + publicSuffixesConditionQuery;
 
         String domainsNotEmpty = FIELD_DOMAIN + ":['' TO *]";
-        String domainsConditionQuery = toMultipleConditionsQuery(originalDomains, FIELD_DOMAIN);
+        String domainsConditionQuery = toMultipleConditionsQueryWithPreCondition(originalDomains, FIELD_DOMAIN);
         String domainsQuery = EXCLUDE_MARKER_TAG + domainsNotEmpty + domainsConditionQuery;
 
-        List<String> hostsToSearch = getActualHostsFromCollections(collectionIds, searchLocation, textToSearch);
-        String originalCollectionsNotEmpty = FIELD_HOST + ":['' TO *]";
-        String originalCollectionsConditionQuery = toMultipleConditionsQuery(hostsToSearch, FIELD_HOST);
-        String originalCollectionsQuery = EXCLUDE_MARKER_TAG + originalCollectionsNotEmpty + originalCollectionsConditionQuery;
+        String collectionsQuery = "";
+        if(collections.size() > 0){
+            String collectionsConditionQuery = toMultipleConditionsQuery(collections, FIELD_COLLECTION);
+            collectionsQuery = EXCLUDE_MARKER_TAG + collectionsConditionQuery;
+        }
 
         String queryString = toEncoded(searchQuery) +
                 "&sort=" + toEncoded(sortByQuery) +
@@ -241,7 +155,7 @@ public class SolrSearchService {
                 "&fq=" + toEncoded(contentTypeQuery) +
                 "&fq=" + toEncoded(publicSuffixesQuery) +
                 "&fq=" + toEncoded(domainsQuery) +
-                "&fq=" + toEncoded(originalCollectionsQuery) +
+                "&fq=" + toEncoded(collectionsQuery) +
                 "&facet.range.gap=%2B1YEAR" +
                 "&facet.range=" + toEncoded(EXCLUDE_POINT_TAG) + FIELD_CRAWL_DATE +
                 "&f." + FIELD_CRAWL_DATE + ".facet.range.start=" + FACET_RANGE_START_YEAR + DATE_PART_AFTER_YEAR +
@@ -253,7 +167,7 @@ public class SolrSearchService {
                 "&fl=" + FIELD_ID + "," + FIELD_TITLE + "," + FIELD_CRAWL_DATE + "," + FIELD_URL + ","
                 + FIELD_WAYBACK_DATE + "," + FIELD_DOMAIN + "," + FIELD_ACCESS_TERMS;
 
-        String[] facets = {FIELD_PUBLIC_SUFFIX, FIELD_CONTENT_TYPE_NORM, FIELD_HOST, FIELD_DOMAIN};
+        String[] facets = {FIELD_PUBLIC_SUFFIX, FIELD_CONTENT_TYPE_NORM, FIELD_HOST, FIELD_DOMAIN, FIELD_COLLECTION};
         return sendRequest(queryString, ContentInfo.class, rowsLimit, startFrom, facets);
     }
 
@@ -277,52 +191,6 @@ public class SolrSearchService {
 
         return dateQuery;
     }
-
-    private List<String> getActualHostsFromCollections(List<String> collectionIds, SearchByEnum searchLocation, String textToSearch) {
-        log.info("Getting list of actual hosts that are contains in collections: " + collectionIds);
-
-        String searchQuery = searchLocation.getSolrSearchLocation() + ":\"" + escapeQueryChars(textToSearch) + "\"";
-        String queryString = toEncoded(searchQuery);
-
-        SolrSearchResult<ContentInfo> result = sendRequest(queryString, ContentInfo.class, 0, 0, FIELD_HOST);
-        LinkedList<String> hostsPairs = result == null || result.getFacetCounts() == null
-                ? new LinkedList<>()
-                : result.getFacetCounts().getFields().getHosts();
-        List<String> actualHosts = toActualHosts(hostsPairs);
-
-        Set<String> hostsToSearch = new HashSet<>();
-
-        collectionIds
-                .stream()
-                .forEach(id -> {
-                    List<String> hostsFromCollection = collectionIdsToHosts.get(id);
-                    if (hostsFromCollection == null) {
-                        return;
-                    }
-
-                    List<String> actualHostsFromCollection = hostsFromCollection
-                            .stream()
-                            .filter(actualHosts::contains)
-                            .collect(toList());
-
-                    hostsToSearch.addAll(actualHostsFromCollection);
-                });
-
-        return new ArrayList<>(hostsToSearch);
-    }
-
-    private List<String> toActualHosts(List<String> hostsPairs) {
-        List<String> actualHosts = new ArrayList<>();
-        for (int i = 0; i < hostsPairs.size(); i += 2) {
-            if ("0".equals(hostsPairs.get(i + 1))) {
-                continue;
-            }
-
-            actualHosts.add(hostsPairs.get(i));
-        }
-        return actualHosts;
-    }
-
 
     private <T extends BodyDocsType> SolrSearchResult<T> sendRequest(String queryString,
                                                                      Class<T> bodyDocsType,